requirements.txt
Django==2.2.13
djangorestframework==3.9.4
requests
